+++ app/Jobs/Integration/SyncEventJob.php
@@ -5,12 +5,14 @@ declare(strict_types=1);
 namespace App\Jobs\Integration;
 
 use App\Models\Event;
+use App\Models\Integration;
 use App\Services\Integration\EventIntegrationService;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Queue\SerializesModels;
+use Illuminate\Support\Facades\Log;
 
 class SyncEventJob implements ShouldQueue
 {
@@ -26,11 +28,18 @@ class SyncEventJob implements ShouldQueue
 
     public function handle(EventIntegrationService $service): void
     {
-        match ($this->action) {
-            'create' => $service->create($this->event),
-            'update' => $service->update($this->event),
-            'delete' => $service->destroy($this->event),
-            default  => throw new \InvalidArgumentException("Unknown action: {$this->action}"),
-        };
+        $integration = Integration::firstOrCreate();
+
+        if ((bool) $integration->status === true) {
+            match ($this->action) {
+                'create' => $service->create($this->event),
+                'update' => $service->update($this->event),
+                'delete' => $service->destroy($this->event),
+                default  => throw new \InvalidArgumentException("Unknown action: {$this->action}"),
+            };
+        } else {
+
+            Log::channel('integration')->error("[Eventicious] Integration is disabled");
+        }
     }
 }
